refactor(macos): extract TimestampDivider and ChatQueueSummaryView from ChatView
